Extract GraknRunnerBase
